Create seperate test fiels for users index, users show, post index, and post show routes
